+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Fixed various scrolling issues in timeline
+- Fixed outbox loading in timeline
+
 ### Security
 
 ## 3.0.2